+++ apps/web/app/(use-page-wrapper)/(main-nav)/bookings/[status]/page.tsx
@@ -6,9 +6,8 @@ import { redirect } from "next/navigation";
 import { z } from "zod";
 
 import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
-import { FeaturesRepository } from "@calcom/features/flags/features.repository";
+import { getUserFeatureRepository } from "@calcom/features/di/containers/UserFeatureRepository";
 import { PermissionCheckService } from "@calcom/features/pbac/services/permission-check.service";
-import { prisma } from "@calcom/prisma";
 import { MembershipRole } from "@calcom/prisma/enums";
 
 import { buildLegacyRequest } from "@lib/buildLegacyCtx";
@@ -54,13 +53,13 @@ const Page = async ({ params }: PageProps) => {
     canReadOthersBookings = teamIdsWithPermission.length > 0;
   }
 
-  const featuresRepository = new FeaturesRepository(prisma);
-  const featureFlags = session?.user?.id
-    ? await featuresRepository.getUserFeaturesStatus(session.user.id, ["bookings-v3", "booking-audit"])
-    : { "bookings-v3": false, "booking-audit": false };
-
-  const bookingsV3Enabled = featureFlags["bookings-v3"] ?? false;
-  const bookingAuditEnabled = featureFlags["booking-audit"] ?? false;
+  const userFeatureRepository = getUserFeatureRepository();
+  const [bookingsV3Enabled, bookingAuditEnabled] = session?.user?.id
+    ? await Promise.all([
+        userFeatureRepository.checkIfUserHasFeature(session.user.id, "bookings-v3"),
+        userFeatureRepository.checkIfUserHasFeature(session.user.id, "booking-audit"),
+      ])
+    : [false, false];
 
   return (
     <ShellMainAppDir
