+++ packages/features/flags/repositories/PrismaTeamFeatureRepository.ts
@@ -1,6 +1,10 @@
+import { captureException } from "@sentry/nextjs";
+
 import type { TeamFeaturesDto } from "@calcom/lib/dto/TeamFeaturesDto";
 import type { PrismaClient } from "@calcom/prisma/client";
-import type { FeatureId } from "../config";
+import { Prisma } from "@calcom/prisma/client";
+
+import type { FeatureId, TeamFeatures } from "../config";
 
 export interface ITeamFeatureRepository {
   findByTeamIdAndFeatureId(teamId: number, featureId: FeatureId): Promise<TeamFeaturesDto | null>;
@@ -18,6 +22,8 @@ export interface ITeamFeatureRepository {
   findAutoOptInByTeamId(teamId: number): Promise<boolean>;
   findAutoOptInByTeamIds(teamIds: number[]): Promise<Record<number, boolean>>;
   setAutoOptIn(teamId: number, enabled: boolean): Promise<void>;
+  checkIfTeamHasFeature(teamId: number, featureId: FeatureId): Promise<boolean>;
+  getEnabledFeatures(teamId: number): Promise<TeamFeatures | null>;
 }
 
 export class PrismaTeamFeatureRepository implements ITeamFeatureRepository {
@@ -159,5 +165,89 @@ export class PrismaTeamFeatureRepository implements ITeamFeatureRepository {
       where: { id: teamId },
       data: { autoOptInFeatures: enabled },
     });
+  }
+
+  /**
+   * Checks if a team or any of its ancestors has access to a specific feature.
+   * Uses a recursive CTE raw SQL query for performance.
+   * Uses tri-state semantics: only treats as enabled if TeamFeatures row exists AND enabled=true.
+   */
+  async checkIfTeamHasFeature(teamId: number, featureId: FeatureId): Promise<boolean> {
+    try {
+      const teamFeature = await this.prisma.teamFeatures.findUnique({
+        where: {
+          teamId_featureId: {
+            teamId,
+            featureId,
+          },
+        },
+        select: { enabled: true },
+      });
+      if (teamFeature) return teamFeature.enabled;
+
+      const query = Prisma.sql`
+        WITH RECURSIVE TeamHierarchy AS (
+          SELECT id, "parentId",
+            CASE WHEN EXISTS (
+              SELECT 1 FROM "TeamFeatures" tf
+              WHERE tf."teamId" = id AND tf."featureId" = ${featureId} AND tf."enabled" = true
+            ) THEN true ELSE false END as has_feature
+          FROM "Team"
+          WHERE id = ${teamId}
+
+          UNION ALL
+
+          SELECT p.id, p."parentId",
+            CASE WHEN EXISTS (
+              SELECT 1 FROM "TeamFeatures" tf
+              WHERE tf."teamId" = p.id AND tf."featureId" = ${featureId} AND tf."enabled" = true
+            ) THEN true ELSE false END as has_feature
+          FROM "Team" p
+          INNER JOIN TeamHierarchy c ON p.id = c."parentId"
+          WHERE NOT c.has_feature
+        )
+        SELECT 1
+        FROM TeamHierarchy
+        WHERE has_feature = true
+        LIMIT 1;
+      `;
+
+      const result = await this.prisma.$queryRaw<unknown[]>(query);
+      return result.length > 0;
+    } catch (err) {
+      captureException(err);
+      console.error(
+        `Recursive feature check failed for team ${teamId}, feature ${featureId}:`,
+        err instanceof Error ? err.message : err
+      );
+      throw err;
+    }
+  }
+
+  /**
+   * Gets all features enabled for a specific team in a map format.
+   */
+  async getEnabledFeatures(teamId: number): Promise<TeamFeatures | null> {
+    const result = await this.prisma.teamFeatures.findMany({
+      where: {
+        teamId,
+        enabled: true,
+      },
+      select: {
+        feature: {
+          select: {
+            slug: true,
+          },
+        },
+      },
+    });
+
+    if (!result.length) return null;
+
+    const features: TeamFeatures = Object.fromEntries(
+      result.map((teamFeature) => [teamFeature.feature.slug, true])
+    ) as TeamFeatures;
+
+    return features;
   }
 }
